Move header to mapbox/ namespace
