+++ include/mapbox/earcut.hpp
